docs(inbox): describe the packages, document public types, drop stale claims
Четвёртый раунд ревью, гигиена перед публикацией.

ОПИСАНИЕ ПАКЕТОВ. Три новых nuspec несли плейсхолдер «Package Description» —
первое, что увидел бы потребитель на nuget.org. Задан Description. Остальные
пакеты репозитория не трогал: они уже отгружены.

ДОКУМЕНТАЦИЯ ТИПОВ. 13 публичных типов не имели summary на самом типе (члены
были документированы). Дописано.

СПЕКУЛЯТИВНЫЙ API. Удалён UseDefaultStrategy(): его никто не вызывает, он
no-op (поле уже инициализировано той же стратегией), и у Outbox аналога нет.
Публичный член это обещание на годы.

КОММЕНТАРИИ, ГОВОРЯЩИЕ С РЕВЬЮЕРОМ. Три комментария объясняли, «как было
раньше», хотя никакого «раньше» у новой библиотеки нет: после мерджа они
стали бы ложью в чужой голове. Один вдобавок врал дважды — ссылался на README,
тогда как требование живёт в CLAUDE.md, относится к Outbox и действует до сих
пор. Убран и устаревший XML-remark, обещавший проверку кавычки в
дискриминаторе: прошлый раунд снял её вместе со всем белым списком.

ФИКСТУРЫ. InboxOptionsValidationTests поднимал реальную базу на каждый из 15
тестов, а нужна она одному. Тест старта хоста вынесен в свою фикстуру,
валидаторы больше не платят за БД. Эталон (OutboxHandlerOptionsValidatorTests)
устроен так же.

InitDelayRange продублирован с Outbox: обе копии помечены комментарием-якорем.
Ссылка инбокса на аутбокс связала бы независимые пакеты, а вынос в
Dex.Cap.Common сменил бы namespace публичного типа уже опубликованного
Dex.Cap.Outbox.AspNetScheduler 8.3.1. Решение за мейнтейнерами.

Тесты InitDelayRange сохранены намеренно: они покрывают ТИП ИНБОКСА, а не
одноимённый тип аутбокса в другой сборке. Удалить их значило бы оставить без
покрытия ровно ту копию, которая может разойтись.
