+++ src/Dex.Cap/Dex.Cap.Inbox.AspNetScheduler/BackgroundServices/InboxRegistryWarmupService.cs
@@ -10,9 +10,9 @@ namespace Dex.Cap.Inbox.AspNetScheduler.BackgroundServices;
 /// </summary>
 /// <remarks>
 /// Реестр строится по сборкам рефлексией и может обнаружить ошибку конфигурации: один дискриминатор у
-/// двух типов, пустой дискриминатор, кавычка в дискриминаторе. Без этого сервиса реестр строился бы
-/// лениво, уже внутри фонового обработчика, где цикл перехватывает исключение и пишет LogCritical:
-/// хост считался бы поднятым, а инбокс молча не обрабатывал бы ничего.
+/// двух типов или пустой дискриминатор. Без этого сервиса реестр строился бы лениво, уже внутри фонового
+/// обработчика, где цикл перехватывает исключение и пишет LogCritical: хост считался бы поднятым, а
+/// инбокс молча не обрабатывал бы ничего.
 /// <para>
 /// Это <see cref="IHostedService"/>, а не <see cref="BackgroundService"/>: работа выполняется в
 /// <see cref="StartAsync"/> синхронно, поэтому исключение роняет старт хоста, как и задумано.
