+++ src/Dex.Cap/Dex.Cap.Inbox.AspNetScheduler/Dex.Cap.Inbox.AspNetScheduler.csproj
@@ -16,6 +16,7 @@
         <PackageProjectUrl>https://github.com/dex-it/dex-common/tree/main/src/Dex.Cap</PackageProjectUrl>
         <PackageLicenseExpression>MIT</PackageLicenseExpression>
         <Authors>dex team</Authors>
+        <Description>Фоновый обработчик, чистильщик и health check для Dex.Cap.Inbox.</Description>
         <ContinuousIntegrationBuild>true</ContinuousIntegrationBuild>
         <GenerateDocumentationFile>true</GenerateDocumentationFile>
         <PublishRepositoryUrl>true</PublishRepositoryUrl>
