+++ src/Dex.Cap/Dex.Cap.Inbox/Exceptions/InboxException.cs
@@ -2,6 +2,9 @@ using System;
 
 namespace Dex.Cap.Inbox.Exceptions;
 
+/// <summary>
+/// Базовый тип ошибок инбокса. Наследники уточняют причину.
+/// </summary>
 public class InboxException : Exception
 {
     public InboxException()
