+++ src/Dex.Cap/Dex.Cap.Inbox/Interfaces/IInboxRetryStrategy.cs
@@ -3,7 +3,13 @@ using Dex.Cap.Inbox.Options;
 
 namespace Dex.Cap.Inbox.Interfaces;
 
+/// <summary>
+/// Когда повторить сообщение, обработка которого завершилась ошибкой.
+/// </summary>
 public interface IInboxRetryStrategy
 {
+    /// <summary>
+    /// Вычислить момент следующей попытки.
+    /// </summary>
     DateTime CalculateNextStartDate(InboxRetryStrategyOptions? options = default);
 }
